change STEP_SIZE from int to long

+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/conf/ConfigDescriptor.java
@@ -377,7 +377,8 @@ public class ConfigDescriptor {
             Boolean.parseBoolean(
                 properties.getProperty("IS_RECENT_QUERY", config.isIS_RECENT_QUERY() + "")));
         config.setSTEP_SIZE(
-            Integer.parseInt(properties.getProperty("STEP_SIZE", config.getSTEP_SIZE() + "")));
+                Long.parseLong(
+                        properties.getProperty("STEP_SIZE",config.getSTEP_SIZE()+ "")));
         config.setOPERATION_PROPORTION(
             properties.getProperty("OPERATION_PROPORTION", config.getOPERATION_PROPORTION()));
         config.setQUERY_SENSOR_NUM(
